refactor: clean up import paths and consolidate utils
