+++ app/search/page.tsx
@@ -1,12 +1,12 @@
 import { ThemeProvider } from '@/components/ThemeProvider';
 import Navigation from '@/components/Navigation';
 import Footer from '@/components/Footer';
-import Section from '@/src/components/kit/Section';
-import PageHeader from '@/src/components/kit/PageHeader';
-import { searchBlogs, fetchAllCategoriesMap, fetchAllSubcategoriesMap } from '@/src/lib/queries';
-import { toInt, pageCount } from '@/src/lib/pagination';
-import Pagination from '@/src/components/pagination';
-import EmptyState from '@/src/components/empty-state';
+import Section from '@/components/kit/Section';
+import PageHeader from '@/components/kit/PageHeader';
+import { searchBlogs, fetchAllCategoriesMap, fetchAllSubcategoriesMap } from '@/lib/queries';
+import { toInt, pageCount } from '@/lib/pagination';
+import Pagination from '@/components/pagination';
+import EmptyState from '@/components/empty-state';
 import Link from 'next/link';
 import SearchForm from './search-form';
