+++ src/components/kit/Section.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { cn } from '@/lib/utils';
 import { motion } from 'framer-motion';
-import { container } from '@/src/lib/motion';
+import { container } from '@/lib/motion';
 import { PropsWithChildren } from 'react';
 
 type SectionProps = PropsWithChildren<{
