+++ src/hooks/use-toast.ts
@@ -126,7 +126,7 @@ export const reducer = (state: State, action: Action): State => {
   }
 };
 
-const listeners: Array<(state: State) => void> = [];
+const listeners: Array<React.Dispatch<React.SetStateAction<State>>> = [];
 
 let memoryState: State = { toasts: [] };
